# claroz-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A RESTful API for the Claroz social platform",
        "title": "Claroz API",
        "termsOfService": "https://claroz.com/terms",
        "contact": {
            "name": "API Support",
            "url": "https://claroz.com/support",
            "email": "lucas.little@claroz.com"
        },
        "license": {
            "name": "Proprietary",
            "url": "https://claroz.com/license"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "schemes": ["http"],
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Authenticate user with email and password",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "tags": ["auth"],
                "summary": "Login user",
                "parameters": [{
                    "description": "Login credentials",
                    "name": "request",
                    "in": "body",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/handlers.LoginRequest"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.AuthResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Register a new user with username, email, and password",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "tags": ["auth"],
                "summary": "Register a new user",
                "parameters": [{
                    "description": "Registration details",
                    "name": "request",
                    "in": "body",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/handlers.RegisterRequest"
                    }
                }],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.AuthResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input or email already registered",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/federation/resolve/{handle}": {
            "get": {
                "description": "Resolves a remote profile from a federated handle (e.g. user.bsky.social)",
                "produces": ["application/json"],
                "tags": ["federation"],
                "summary": "Resolve a remote profile by handle",
                "parameters": [{
                    "type": "string",
                    "description": "Remote handle to resolve",
                    "name": "handle",
                    "in": "path",
                    "required": true
                }],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                }
            }
        },
        "/federation/sync/{did}": {
            "post": {
                "description": "Syncs the latest data for a remote profile",
                "produces": ["application/json"],
                "tags": ["federation"],
                "summary": "Sync a remote profile",
                "parameters": [{
                    "type": "string",
                    "description": "DID of the remote profile",
                    "name": "did",
                    "in": "path",
                    "required": true
                }],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.LoginRequest": {
            "type": "object",
            "required": ["email", "password"],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "lucas.little@claroz.com"
                },
                "password": {
                    "type": "string",
                    "example": "password123"
                }
            }
        },
        "handlers.RegisterRequest": {
            "type": "object",
            "required": ["username", "email", "password"],
            "properties": {
                "username": {
                    "type": "string",
                    "example": "lucas"
                },
                "email": {
                    "type": "string",
                    "example": "lucas.little@claroz.com"
                },
                "password": {
                    "type": "string",
                    "example": "password123"
                },
                "full_name": {
                    "type": "string",
                    "example": "Lucas Little"
                }
            }
        },
        "handlers.AuthResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "user": {
                    "$ref": "#/definitions/models.User"
                }
            }
        },
...
        "models.User": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "username": {
                    "type": "string",
                    "example": "lucas"
                },
                "email": {
                    "type": "string",
                    "example": "lucas.little@claroz.com"
                },
                "full_name": {
                    "type": "string",
                    "example": "Lucas Little"
                },
                "bio": {
                    "type": "string",
                    "example": "Software engineer at Claroz"
                },
                "avatar": {
                    "type": "string",
                    "example": "https://claroz.com/avatars/lucas.jpg"
                },
                "did": {
                    "type": "string",
                    "example": "did:web:claroz.com:lucas"
                },
                "handle": {
                    "type": "string",
                    "example": "@lucas"
                },
                "federation_type": {
                    "type": "string",
                    "example": "local"
                },
                "created_at": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2024-01-26T00:35:27Z"
                },
                "updated_at": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2024-01-26T00:35:27Z"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header",
            "description": "Type \"Bearer\" followed by a space and JWT token."
        }
    }
}